requirements.txt
python-telegram-bot==8.1.1
ruamel.yaml==0.15.34
peewee==2.10.2
Pillow==4.3.0
ImageHash == 3.7
requests==2.18.4
pydantic==0.5